+++ packages/headless/src/app/headless-engine.ts
@@ -186,7 +186,9 @@ export interface Engine<State = SearchAppState> {
    */
   store: Store;
   renewAccessToken: () => Promise<string>;
-  /** The logging instance used by headless */
+  /**
+   * The logger instance used by headless.
+   * */
   logger: Logger;
 }