+++ packages/headless/src/features/search-parameters/search-parameter-actions.ts
@@ -1,12 +1,23 @@
 import {createAction} from '@reduxjs/toolkit';
-import {
-  EnableQuerySyntaxParam,
-  QueryParam,
-} from '../../api/search/search-api-params';
+import {SearchRequest} from '../../api/search/search/search-request';
 import {validatePayloadSchema} from '../../utils/validate-payload';
 import {searchParametersDefinition} from './search-parameter-schema';
 
-export type SearchParameters = QueryParam & EnableQuerySyntaxParam;
+export type SearchParameters = Omit<
+  SearchRequest,
+  | 'organizationId'
+  | 'accessToken'
+  | 'visitorId'
+  | 'url'
+  | 'enableDidYouMean'
+  | 'fieldsToInclude'
+  | 'facetOptions'
+  | 'facets'
+  | 'searchHub'
+  | 'pipeline'
+  | 'context'
+  | 'debug'
+>;
 
 /** Restores search parameters from e.g. a url*/
 export const restoreSearchParameters = createAction(